Converted projects to use so-called "minimal hosting" with all initialization inside Program.cs instead of using a Startup class.
