+++ fundamentals/streaming/DynamicBinaryStream/Program.cs
@@ -1,5 +1,13 @@
-using Microsoft.AspNetCore;
+using System;
+
+using ASPNetCoreStreamingExample.DynamicBinaryStream.Model;
+
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Hosting.Server.Features;
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace ASPNetCoreStreamingExample.DynamicBinaryStream
 {
@@ -7,10 +15,44 @@ namespace ASPNetCoreStreamingExample.DynamicBinaryStream
     {
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
-        }
+            var builder = WebApplication.CreateBuilder(args);
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-          => WebHost.CreateDefaultBuilder(args).UseStartup<Startup>();
+            // Song lyrics source to be injected into instances.
+            builder.Services.AddSingleton<ILyricsSource, LyricsSource>();
+
+            // Frame generator for animation
+            builder.Services.AddSingleton<IFrameSourceAsync, FrameSourceAsync>();
+
+            builder.Services.AddControllers();
+
+            var app = builder.Build();
+
+            if (app.Environment.IsDevelopment())
+                app.UseDeveloperExceptionPage();
+
+            app.UseRouting();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints => endpoints.MapControllers());
+
+            var appBuilder = (IApplicationBuilder)app;
+
+            var hostLifetime = appBuilder.ApplicationServices.GetRequiredService<IHostApplicationLifetime>();
+
+            hostLifetime.ApplicationStarted.Register(
+                () =>
+                {
+                    var addressesFeature = appBuilder.ServerFeatures.Get<IServerAddressesFeature>()!;
+
+                    Console.WriteLine();
+                    Console.WriteLine("Please browse to the hosted service:");
+
+                    foreach (var address in addressesFeature.Addresses)
+                        Console.WriteLine("* {0}/menu", address.Replace("0.0.0.0", "localhost"));
+
+                    Console.WriteLine();
+                });
+
+            app.Run();
+        }
     }
 }
